Add online status to user
I'm not sure if there is a cleaner way to do this from the model, if it would be
better to just augment the data after I convert it to a plain object, or if
online should be it's own endpoint.

In the grand scheme, it should probably be pushed on change instead of queried,
but that seems a bit out of scope right now.